All Functions in Character Selection work now.
Create Delete Character 100%
Name Check 90%
SelectServer 100%
Select Characters 100%
Start Game Button 100%

+++ SunCommon/Entities/Inventory.cs
@@ -9,8 +9,8 @@ namespace SunCommon.Entities
     public class Inventory
     {
         public int Id { get; set; }
-        public int Money { get; set; }
-        public int InventoryLock { get; set; }
+        public long Money { get; set; }
+        public byte InventoryLock { get; set; }
         public byte[] InventoryItem { get; set; }
         public byte[] TmpInventoryItem { get; set; }
         public byte[] EquipItem { get; set; }
